requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dateutil==2.8.2
requests==2.32.5
pandas==2.3.3
yfinance==0.2.32
python-dotenv==1.0.0
beautifulsoup4==4.14.3
selenium==4.15.2
webdriver-manager==4.0.1
